+++ src/main.js
@@ -1,7 +1,6 @@
 'use strict';
 
-require('./check.js');
 require('./resizer.js');
-require('./upload-stat.min.js');
 require('./upload.js');
-require('./pictures.js');
+require('./pictures/pictures');
+// require('./pictures/filters');
